==> fire-prediction-notebooks/8_create_eco_stats.py <==
# -*- coding: utf-8 -*-
"""8-create_eco_stats.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1djFtf9_9WIiV5GMLg-B-X0IEqNasqa2o

Input:

1.   Eco Info File
2.   AL_dataset.csv

Output:
AL_eco_zone_stats.csv.gz
"""

from google.colab import files, drive
import pandas as pd
import numpy as np

drive.mount('/content/gdrive')
root_path = '/content/gdrive/My Drive/Capstone Public Folder/Data/'
provinces ={"MB", "SK", "AL", "ON", "QB", "BC"}
province = "CA"

ecoInfo = pd.read_csv(root_path +"Colab Data/EcoZoneNumberNameMapping.csv.gz", compression = "gzip")

stats = []

for pr in provinces:
  stats.append(pd.read_csv(root_path  f'{pr} Location/{pr}_dataset.csv.gz', compression = "gzip"))

stats_df = pd.concat(stats)

drop_columns = {
    'LOCATION_KEY', 'MAX_TEMP', 'MIN_TEMP', 'MEAN_TEMP', 'HEAT_DEG_DAYS',
       'COOL_DEG_DAYS', 'TOTAL_RAIN', 'TOTAL_SNOW', 'TOTAL_PRECIP',
       'SNOW_ON_GRND', 'DIR_OF_MAX_GUST', 'SPD_OF_MAX_GUST', 'TEMP_12_4',
       'DEW_POINT_TEMP_12_4', 'REL_HUM_12_4', 'LOCATION_DATE_KEY', 'FIRE',
       'LATITUDE', 'LONGITUDE', 'ECOREGION', 'ECODISTRICT'
}

for i in drop_columns: 
  del stats_df[i]

stats.fillna(0)

df = stats_df.groupby('ECOZONE', as_index = False).agg({'TOTAL_SIZE_HA_OLD': 'sum', 'TOTAL_DURATION_OLD': 'sum', 'AVERAGE_SIZE_HA_OLD': 'mean', 'AVERAGE_DURATION_OLD': 'mean'})

merged_df = pd.merge(ecoInfo, df, how = 'left', left_on = 'ECOZONE', right_on = 'ECOZONE')

merged_df.to_csv(root_path + province + " Location/"+ province + '_eco_zone_stats.csv.gz', compression = 'gzip', index = False)
